2nd try

+++ Horse.java
@@ -3,10 +3,10 @@
 public class Horse{
 
   String name;
-  int weight;
+  double weight;
   boolean tame;
 
-    public Horse(String name, int weight, boolean tame){
+    public Horse(String name, double weight, boolean tame){
 
       this.name= name;
       this.weight= weight;
